requirements.txt
boto3==1.35.34
django-storages==1.14.4